+++ Sources/KSCrash/Recording/Tools/Raygun_KSDate.h
@@ -1,5 +1,5 @@
 //
-//  KSDate.c
+// KSDate.h
 //
 // Copyright 2016 Karl Stenerud.
 //
@@ -22,19 +22,26 @@
 // THE SOFTWARE.
 //
 
-#include "KSDate.h"
-#include <stdio.h>
-#include <time.h>
+#ifndef KSDate_h
+#define KSDate_h
 
-void ksdate_utcStringFromTimestamp(time_t timestamp, char* buffer21Chars)
-{
-    struct tm result = {0};
-    gmtime_r(&timestamp, &result);
-    snprintf(buffer21Chars, 21, "%04d-%02d-%02dT%02d:%02d:%02dZ",
-             result.tm_year + 1900,
-             result.tm_mon+1,
-             result.tm_mday,
-             result.tm_hour,
-             result.tm_min,
-             result.tm_sec);
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+
+#include <sys/types.h>
+
+/** Convert a UNIX timestamp to an RFC3339 string representation.
+ *
+ * @param timestamp The date to convert.
+ *
+ * @param buffer21Chars A buffer of at least 21 chars to hold the RFC3339 date string.
+ */
+void raygun_ksdate_utcStringFromTimestamp(time_t timestamp, char* buffer21Chars);
+
+#ifdef __cplusplus
 }
+#endif
+
+#endif /* KSDate_h */
